+++ cmd/DetectorAppDelegate.m
@@ -4,36 +4,33 @@
 #import <DetectorAppDelegate.h>
 
 void queryWire(void *pathPointer, void *createdAtPointer);
+void ui();
+void openPreferences();
+void openHelp();
+void savePreferences();
 
 @implementation DetectorAppDelegate
 @synthesize queryResults;
 
 - (void)applicationDidFinishLaunching:(NSNotification *)aNotification
 {
+    ui();
+
     query = [[NSMetadataQuery alloc] init];
     NSNotificationCenter* defaultCenter = [NSNotificationCenter defaultCenter];
 
-    [
-        defaultCenter
-        addObserver:self
-        selector:@selector(queryUpdated:)
-        name:NSMetadataQueryDidStartGatheringNotification
-        object:query
-    ];
-    [
-        defaultCenter
-        addObserver:self
-        selector:@selector(queryUpdated:)
-        name:NSMetadataQueryDidUpdateNotification
-        object:query
-    ];
-    [
-        defaultCenter
-        addObserver:self
-        selector:@selector(queryUpdated:)
-        name:NSMetadataQueryDidFinishGatheringNotification
-        object:query
-    ];
+    [defaultCenter addObserver:self
+                      selector:@selector(queryUpdated:)
+                          name:NSMetadataQueryDidStartGatheringNotification
+                        object:query];
+    [defaultCenter addObserver:self
+                      selector:@selector(queryUpdated:)
+                          name:NSMetadataQueryDidUpdateNotification
+                        object:query];
+    [defaultCenter addObserver:self
+                      selector:@selector(queryUpdated:)
+                          name:NSMetadataQueryDidFinishGatheringNotification
+                        object:query];
 
     [query setDelegate:self];
     [query setPredicate:[NSPredicate predicateWithFormat:@"kMDItemIsScreenCapture = 1"]];
@@ -66,9 +63,8 @@ void queryWire(void *pathPointer, void *createdAtPointer);
 
 - (NSMutableArray *)accessToFolder {
     const char *home = getpwuid(getuid())->pw_dir;
-    NSString *path = [[NSFileManager defaultManager]
-                      stringWithFileSystemRepresentation:home
-                      length:strlen(home)];
+    NSString *path = [[NSFileManager defaultManager] stringWithFileSystemRepresentation:home
+                                                                                 length:strlen(home)];
     NSString *realHomeDirectory = [[NSURL fileURLWithPath:path isDirectory:YES] path];
 
     NSMutableArray *pathURLs = [NSMutableArray array];
@@ -77,5 +73,17 @@ void queryWire(void *pathPointer, void *createdAtPointer);
     [pathURLs addObject:[NSURL fileURLWithPath:[NSString stringWithFormat:@"%@/Pictures", realHomeDirectory] isDirectory:true]];
 
     return pathURLs;
+}
+
+- (void)openPreferencesSel {
+    openPreferences();
+}
+
+- (void)openHelpSel {
+    openHelp();
+}
+
+- (void)savePreferencesSel {
+    savePreferences();
 }
 @end